Currency Test Update; Invalid currency fails to evaluate
Before: When trying to evaluate an unknown currency using CountCurrencyEvaluatableValue, an exception was thrown if the targeted currency doesn't exist.
Fix: Failing to identify the currency will fail the evaluation, without throwing an exception.
Future: The database methods for getting things directly by id should be private. There shouldn't be calls to it from otuside the database. Will note this for later.

ClickUp #868arcyut