Tokenization Starting Tests
Created a new Tests class, TokenizationStartingTests. This will work as a sort of test-driven-diary of features as we're re-implementing them. This pattern will probably continue to be common throughout development.

For now, we're testing if we can upload a command and get it back (check) and if we can send text in and get the expected parsing (not check). The actual logic for translating text to concepts isn't there yet.